refactor: Rename main package to checkstyle
