+++ setup.cfg
@@ -22,4 +22,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    checkstyle = main.cli:main
+    checkstyle = checkstyle.cli:main
