--- src/__main__.py
@@ -1,4 +1,0 @@
-from main.cli import main
-
-if __name__ == "__main__":
-    main()
