+++ src/checkstyle/utils.py
